refactorisation colors et size pour les icones
