+++ frontend/src/components/color.js
@@ -1,0 +1,16 @@
+const color = {
+	grey: "#EBE8E8",
+	oldpink: "#F0DBD8",
+	salmon: "#FFCAB8",
+	gold: "#F9C153",
+	purple: "#241F5D",
+	velvet: "#5353A2",
+	plum: "#6767AD",
+	emerald: "#78C4D2",
+	turquoise: "#7ED0DF",
+	asura: "#94DAD5",
+	orange: "#FF8B66",
+	brick: "#EB5933",
+	white: "#FFFFFF",
+};
+export default color;
