+++ frontend/src/components/size.js
@@ -1,0 +1,7 @@
+const size = {
+	small: "1em",
+	medium: "2em",
+	large: "4em",
+	xlarge: "5em",
+};
+export default size;
